+++ shopping-web/src/components/signup/SignUp.jsx
@@ -6,17 +6,13 @@ import { FaGithub } from "react-icons/fa";
 import { useForm } from "react-hook-form";
 import toast from "react-hot-toast";
 import { useMyContext } from "../../store/ContextApi";
+import AddressInput from "../Auth/AddressInput";
 const SignUp = () => {
   const apiUrl = import.meta.env.VITE_APP_API_URL;
   const [role, setRole] = useState();
   const [loading, setLoading] = useState(false);
   const { token } = useMyContext();
   const navigate = useNavigate();
-
-
-
-
-
   
   const {
     register,
@@ -56,17 +52,17 @@ const SignUp = () => {
   const onSubmitHandler = async (data) => {
     console.log("회원가입 요청 데이터:", data); // 로그 확인
   
-    const { username, name, email, phoneNumber, postcode, address, detailAddress, extraAddress, password } = data;
+    const { username, name, email, phoneNumber,  password , postcode, address, detailAddress, extraAddress} = data;
     const sendData = { 
       username, 
       name,
       email, 
       phoneNumber, 
-      postcode, 
-      address, 
-      detailAddress: detailAddress || "",  // null 방지
-      extraAddress: extraAddress || "", 
       password, 
+      postcode,
+      address,
+      detailAddress,
+      extraAddress,
       role: [role] 
     };
 
@@ -100,43 +96,9 @@ const SignUp = () => {
     if (token) navigate("/login");
   }, [navigate, token]);
 
-  useEffect(() => {
-    const script = document.createElement("script");
-    script.src = "//t1.daumcdn.net/mapjsapi/bundle/postcode/prod/postcode.v2.js";
-    script.async = true;
-    document.body.appendChild(script);
-  }, []);
+ 
 
-  const handleAddressSearch = () => {
-    if (!window.daum) {
-      alert("주소 검색 기능을 불러오는 중입니다. 잠시 후 다시 시도해주세요.");
-      return;
-    }
-
-    new window.daum.Postcode({
-      oncomplete: function (data) {
-        let addr = data.roadAddress || data.jibunAddress; // 도로명 주소 또는 지번 주소
-        let extraAddr = "";
-
-        if (data.userSelectedType === "R") {
-          if (data.bname !== "" && /[동|로|가]$/g.test(data.bname)) {
-            extraAddr += data.bname;
-          }
-          if (data.buildingName !== "" && data.apartment === "Y") {
-            extraAddr += (extraAddr !== "" ? `, ${data.buildingName}` : data.buildingName);
-          }
-          if (extraAddr !== "") {
-            extraAddr = `(${extraAddr})`;
-          }
-        }
-
-        setValue("postcode", data.zonecode); // 우편번호
-        setValue("address", addr); // 주소
-        setValue("extraAddress", extraAddr); // 참고 항목
-        setValue("detailAddress", ""); // 상세 주소 초기화
-      },
-    }).open();
-  };
+ 
 
   return (
     <div className="flex h-screen bg-white mt-30 mb-30">
@@ -205,38 +167,9 @@ const SignUp = () => {
               })}
             />
             {errors.email && <p className="text-red-500 text-sm">{errors.email.message}</p>}
-
-          {/* 우편번호 + 우편번호 찾기 버튼 */}
-          <div className="flex gap-2">
-              <input type="text" className="w-2/3 p-2 border rounded-md" placeholder="우편번호" {...register("postcode")} readOnly />
-              <button
-            type="button"
-            onClick={handleAddressSearch}
-            className="flex gap-2 items-center justify-center flex-1 border p-3 shadow-sm shadow-gray-200 rounded-md hover:bg-gray-300 transition-all duration-300 w-full"
-          >
-            <span className="font-semibold text-sm">우편번호 찾기</span>
-          </button>
-            </div>
-
-            {/* 주소 */}
-            <input type="text" className="w-full p-2 border rounded-md" placeholder="주소" {...register("address")} readOnly />
-
-            {/* 상세 주소 + 참고 항목 */}
-            <div className="flex gap-2">
-            <input
-              type="text"
-              className="w-2/4 p-2 border rounded-md"
-              placeholder="상세주소"
-              {...register("detailAddress", {
-                required: "상세 주소를 입력해주세요",
-                onChange: (e) => setValue("detailAddress", e.target.value),
-              })}
-            />
-              <input type="text" className="w-2/4 p-2 border rounded-md" placeholder="추가 주소" {...register("extraAddress")}  />
-            </div>
-        
-            {errors.address && <p className="text-red-500 text-sm">{errors.address.message}</p>}
-
+              
+            <AddressInput register={register} setValue={setValue}/>
+      
             {/* 핸드폰 번호 */}
             <input
               type="text"
